Solved issue with poor/biased visualisation
The problem of the visualisation seeming to bias only one side (left or right) is solved by increasing the number of nearest neighbours used for the knnsearch in the algo2only_func_fixed_xy, which was too small at 4nn. I had maintained it at that value because previous attempts at changing it resulted in very little change, but now it's the different between poor visualisation and a great one.

Also corrected the normals re-orientation in the main SN_algo2 so that it points "away" from the surface and not into it.